requirements.txt
python-telegram-bot==22.5
aiohttp>=3.8
requests
python-dotenv
psutil
schedule
gunicorn              # اختياري إن احتجت لكن لن تستخدمه إذا ستشغّل python app.py
uvloop               # اختياري لتحسين الأداء في بيئات asyncio